requirements.txt
colorbrewer==0.2.0
folium==0.17.0
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
Pillow==9.0.1
selenium==4.22.0
